Fixed DSPNetwork::findModule, no more duplicate ModuleSamplePlayers
There was a race condition in findModule, also it didn't check the new
item list for the module. This sometimes lead to a situation that there
were multiple ModuleSamplePlayers, if there was a delay when opening the
sound card.

At one point there were over 900 ModuleSamplePlayer instances running at
the same time