Fixed rendering latency issues with glFinish, decreased GPU memory usage.
Of course, this fix isn't free, it might slow down rendering, depending
on the application and the platform. However, lower latency is usually
nicer than better rendering speed.

Also, since after this we are using only one shared buffer, the memory
usage in that is now 25% of what it used to be.

Related to #4096.