Added predictive limiter to sound output. Fixes #7734.
This introduces a bit of latency (currently 30 samples), but that should be insignificant. The current implementation is not
really configurable. Options can be added if necessary.

The limiter is now on all the time, but probably this is what we want. There is no situation where we would want the output to clip.