Changed frame numbers to unsigned int in RenderThread and RenderContext.
This is because the frame number are unsigned ints in GLX_NV_swap_group.
After this change fewer cast are needed and the types are the same in
different places of the code. If we do wrap this 32 bit counter at some
point, then at least all of them wrap roughly at the same time.