Add Radiant::PlatformUtils::setupConsole()
In practice, calling this function will cause stdout, stderr, and stdin in the
calling process to use the ones in the parent terminal. This resolves the issue
on Windows where we don't want to open a new terminal window, but we still want
stdout to appear in a terminal window if we start an application from the
command line.

Related to #14930