Remove JS support, wrapper generator, tests, examples, apps, docs
Most of them do not work anymore, there are no plans to fix them and the
whole approach how to implement it is not correct (anymore).

Additionally new limited scripting support is going to be added to Canvus
UI tests (and maybe Cornerstone later), so let's not have two conflicting
implementations around.