Simplify Punctual::TaskScheduler executor usage
This makes it cleaner to use various executors with Valuable::Events.

The actual architecture didn't change, these still have the same issue
as before; user don't know if there is someone driving these executors,
but that issue is not part of this change, I just wanted to tidy the
notation.