Fix off-by-one in bezier tesselator sharp corner detector
If the tangent at a point is much different from the tangent in the previous
point, there is a sharp corner that is rendered as a round join. Previously
the join was rendered at the new point, even though the sharp corner was
actually at the previous point. With normal annotation data the points are
so close to each other in sharp corners, that this wasn't noticed, but in
generated splines this caused bad rendering artifacts, see the images in
the MR.